Mod Colors
Blends now allows for an offset which when a card has three or more colors, will rotate which sections of the cards get which color (eg. offset:1 causes rwu to render as wbr).

This is set up in a way that it can be easily hooked into by a style.

Moved the literal strings that direct to masks in the magic-identity-new.mse-includes to a seporate text file to clean the look of those lines.