Updated "borderless" style
Updated the "Borderless" style for cards using the M15 Mainframe. Only changes necessary were updates to the mask images, plus changing which stamp mask is used for borderless cards.

The prior-existing style had a number of problems compared to the actual implementation of Borderless cards for "real" MTG. Some of these issues are fixed separately, with the added support for the "inverted" style. The changes in this commit are as follows:

- Added the cutoff bottom frame that is present on all non-planeswalker borderless cards, rather than using the black fade that is present only on borderless planeswalkers
- Added slight transparency to the name, typeline, and rules text areas to match the transparency used on ALL real borderless cards
- Switched the holofoil stamp mask to use the same mask as all other cards, because the area around the holofoil stamp should just be filled with black